requirements.txt
absl-py==0.9.0
appdirs==1.4.3
arrow==0.15.8
astroid==2.4.2
astunparse==1.6.3
attrs==19.3.0
audioread==2.1.8
backcall==0.2.0
binaryornot==0.4.4
bleach==3.1.5
boto3==1.14.36
botocore==1.17.36
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
colorama==0.4.3
contextlib2==0.6.0
cookiecutter==1.7.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
docutils==0.15.2
entrypoints==0.3
fastprogress==0.2.3
Fiona==1.8.13.post1
future==0.18.2
gast==0.3.3
geopandas==0.8.0
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.1
Jinja2==2.11.2
jinja2-time==0.2.0
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.5
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
librosa==0.7.2
llvmlite==0.31.0
lockfile==0.12.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
mistune==0.8.4
msgpack==0.6.2
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numba==0.48.0
numpy==1.19.0
oauthlib==3.1.0
opencv-python==4.3.0.36
opt-einsum==3.2.1
packaging==20.3
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pdfkit==0.6.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkg-resources==0.0.0
plotly==4.8.2
poyo==0.5.0
progress==1.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.6
PyPDF2==1.26.0
pyproj==2.6.1.post1
pyreadr==0.2.9
pyrsistent==0.16.0
python-dateutil==2.8.1
python-slugify==4.0.1
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
retrying==1.3.3
rsa==4.6
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
Shapely==1.7.0
six==1.14.0
sklearn==0.0
SoundFile==0.10.3.post1
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.1
torch==1.5.1
torchvision==0.6.1
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
traittypes==0.2.1
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
